backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
websocket-client==1.8.0
requests==2.32.3
Pillow==11.1.0
